# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: FIS Process API
  version: v1
  description: Complete REST API contract for the FIS ledger service.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /journal-entries:
    post:
      summary: Create manual journal entry
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ActorRoleHeader'
        - $ref: '#/components/parameters/TraceparentHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    get:
      summary: List journal entries
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: postedDateFrom
          in: query
          schema:
            type: string
            format: date
        - name: postedDateTo
          in: query
          schema:
            type: string
            format: date
        - name: accountCode
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PENDING_APPROVAL, POSTED, REVERSAL, CORRECTION, REJECTED]
        - name: referenceId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Journal entries page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryPage'
  /journal-entries/batch:
    post:
      summary: Create manual journal entries in batch
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ActorRoleHeader'
        - $ref: '#/components/parameters/TraceparentHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalEntryBatchRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryBatchResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '409':
          $ref: '#/components/responses/DuplicateIdempotencyKey'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /journal-entries/{id}:
    get:
      summary: Get journal entry by id
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /journal-entries/{id}/reverse:
    post:
      summary: Reverse posted journal entry
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReversalRequest'
      responses:
        '201':
          description: Reversal posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReversalResponse'
        '409':
          $ref: '#/components/responses/InvalidReversal'
  /journal-entries/{id}/correct:
    post:
      summary: Correct posted journal entry (reverse + replacement)
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequest'
      responses:
        '201':
          description: Correction completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReversalResponse'
        '409':
          $ref: '#/components/responses/InvalidReversal'
  /journal-entries/{id}/submit:
    post:
      summary: Submit draft journal entry for approval
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWorkflowRequest'
      responses:
        '200':
          description: Workflow submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalWorkflowActionResponse'
  /journal-entries/{id}/approve:
    post:
      summary: Approve pending workflow journal entry
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ActorRoleHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveWorkflowRequest'
      responses:
        '201':
          description: Workflow approved and posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalWorkflowActionResponse'
  /journal-entries/{id}/reject:
    post:
      summary: Reject pending workflow journal entry
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectWorkflowRequest'
      responses:
        '200':
          description: Workflow rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalWorkflowActionResponse'
...
  /revaluations/periods/{periodId}:
    post:
      summary: Run period-end revaluation
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: periodId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRevaluationRequest'
      responses:
        '201':
          description: Revaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevaluationResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /revaluations/periods/{periodId}/translation:
    post:
      summary: Run functional-currency translation (CTA/OCI) for income statement items
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: periodId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunTranslationRequest'
      responses:
        '201':
          description: Translation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    ActorRoleHeader:
      name: X-Actor-Role
      in: header
      required: false
      schema:
        type: string
...
    TraceparentHeader:
      name: traceparent
      in: header
      required: false
      schema:
        type: string
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Entity not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    DuplicateIdempotencyKey:
      description: Duplicate idempotency key with conflicting payload
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InvalidReversal:
      description: Reversal is invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: Business rule rejection
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimitExceeded:
      description: Too many requests for protected posting endpoints
      headers:
        Retry-After:
          schema:
            type: integer
          description: Retry window in seconds
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    JournalLineRequest:
      type: object
      required: [accountCode, amountCents, isCredit]
      properties:
        accountCode:
          type: string
        amountCents:
          type: integer
          format: int64
        isCredit:
          type: boolean
        dimensions:
          type: object
          additionalProperties:
            type: string
    CreateJournalEntryRequest:
      type: object
      required: [eventId, postedDate, transactionCurrency, createdBy, lines]
      properties:
        eventId:
          type: string
        postedDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
          nullable: true
        transactionDate:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        referenceId:
          type: string
          nullable: true
        transactionCurrency:
          type: string
        createdBy:
          type: string
        autoReverse:
          type: boolean
          default: false
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalLineRequest'
    CreateJournalEntryBatchRequest:
      type: object
      required: [batchMode, entries]
      properties:
        batchMode:
          $ref: '#/components/schemas/JournalBatchMode'
        entries:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateJournalEntryRequest'
    JournalBatchMode:
      type: string
      enum: [POST_NOW, CREATE_DRAFTS]
    JournalEntryResponse:
      type: object
      properties:
        journalEntryId:
          type: string
          format: uuid
        postedDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
        transactionDate:
          type: string
          format: date
        status:
          type: string
        description:
          type: string
          nullable: true
        referenceId:
          type: string
          nullable: true
        transactionCurrency:
          type: string
        baseCurrency:
          type: string
        exchangeRate:
          type: number
        lineCount:
          type: integer
        reversalOfId:
          type: string
          format: uuid
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        fiscalYear:
          type: integer
          nullable: true
        sequenceNumber:
          type: integer
          format: int64
          nullable: true
    JournalEntryBatchResponse:
      type: object
      properties:
        batchMode:
          $ref: '#/components/schemas/JournalBatchMode'
        count:
          type: integer
        entries:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryResponse'
    JournalEntryPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
    ReversalRequest:
      type: object
      required: [eventId, reason, createdBy]
      properties:
        eventId:
          type: string
        reason:
          type: string
        createdBy:
          type: string
    CorrectionRequest:
      type: object
      required: [eventId, reversalEventId, postedDate, transactionCurrency, createdBy, lines]
      properties:
        eventId:
          type: string
        reversalEventId:
          type: string
        postedDate:
          type: string
          format: date
        description:
          type: string
          nullable: true
        referenceId:
          type: string
          nullable: true
        transactionCurrency:
          type: string
        createdBy:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalLineRequest'
    ReversalResponse:
      type: object
      properties:
        reversalJournalEntryId:
          type: string
          format: uuid
        originalJournalEntryId:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string
    SubmitWorkflowRequest:
      type: object
      required: [submittedBy]
      properties:
        submittedBy:
          type: string
    ApproveWorkflowRequest:
      type: object
      required: [approvedBy]
      properties:
        approvedBy:
          type: string
    RejectWorkflowRequest:
      type: object
      required: [rejectedBy, reason]
      properties:
        rejectedBy:
          type: string
        reason:
          type: string
    JournalWorkflowActionResponse:
      type: object
      properties:
        workflowId:
          type: string
          format: uuid
        status:
          type: string
          enum: [DRAFT, PENDING_APPROVAL, APPROVED, REJECTED]
        postedJournalEntryId:
          type: string
          format: uuid
          nullable: true
        message:
          type: string
...
    RunRevaluationRequest:
      type: object
      required: [eventId, createdBy, reserveAccountCode, gainAccountCode, lossAccountCode]
      properties:
        eventId:
          type: string
        createdBy:
          type: string
        reserveAccountCode:
          type: string
        gainAccountCode:
          type: string
        lossAccountCode:
          type: string
    RunTranslationRequest:
      type: object
      required: [eventId, createdBy, ctaOciAccountCode, translationReserveAccountCode]
      properties:
        eventId:
          type: string
        createdBy:
          type: string
        ctaOciAccountCode:
          type: string
        translationReserveAccountCode:
          type: string
    RevaluationResponse:
      type: object
      properties:
        periodId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        status:
          type: string
        generatedJournalEntryIds:
          type: array
          items:
            type: string
            format: uuid
    TranslationResponse:
      type: object
      properties:
        periodId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        status:
          type: string
        generatedJournalEntryIds:
          type: array
          items:
            type: string
            format: uuid
...